Refactor submit function to own file
